requirements.txt
pillow~=10.4.0
rectpack~=0.2.2